--- src/components/cart/user-shopping-cart/user-shopping-cart.page.vue
@@ -1,25 +1,0 @@
-<script setup>
-  import theToolbar from '../../shared/the-toolbar.component.vue';
-  import userCart from './user-shopping-cart.component.vue';
-</script>
-
-<template>
-  <main class="main">
-    <the-toolbar/>
-    <user-cart/>
-  </main>
-</template>
-
-<style scoped>
-  .main {
-    display: flex;
-    justify-content: space-around;
-    min-height: 100vh;
-    margin: auto;
-  }
-  @media (max-width: 600px) {
-    .main {
-      width: 100vw;
-    }
-  }
-</style>
