+++ src/components/client-layout/catalogue/catalogue-shirts-description/shirt-description.component.vue
