+++ src/components/client-layout/catalogue/shirts-catalogue/shirt-filters.component.vue
