+++ src/components/client-layout/shopping-list/user-shopping-list/user-shopping-list.component.vue
