+++ src/components/shared/profile/edit-user-profile/edit-user-password/edit-user-password.page.vue
