+++ src/routes/modules/businessmans.routes.js
@@ -1,4 +1,4 @@
-﻿import BusinessmanSectionPage from "@/components/businessmans/pages/businessman-section.page.vue";
+﻿import BusinessmanSectionPage from "@/components/client-layout/businessmans/pages/businessman-section.page.vue";
 
 export const businessmansRoutes = [
     {path:'/businessman-list', component: BusinessmanSectionPage, meta: { requiresAuth: true, userRole: 'CLIENT' }}
