+++ src/routes/modules/catalogue.routes.js
@@ -1,5 +1,5 @@
-import ShirtsCataloguePage from "@/components/catalogue/shirts-catalogue/shirts-catalogue.page.vue";
-import ShirtDescriptionPage from "@/components/catalogue/catalogue-shirts-description/shirt-description.page.vue";
+import ShirtsCataloguePage from "@/components/client-layout/catalogue/shirts-catalogue/shirts-catalogue.page.vue";
+import ShirtDescriptionPage from "@/components/client-layout/catalogue/catalogue-shirts-description/shirt-description.page.vue";
 
 export const catalogueRoutes = [
   { path: '/catalogue', component: ShirtsCataloguePage},
