+++ src/routes/modules/posts.routes.js
@@ -1,6 +1,6 @@
-import EditShirtPage from "@/components/posts/edit-post/edit-shirt.page.vue";
-import PublishShirtPage from "@/components/posts/publish-post/publish-shirt.page.vue";
-import PostCataloguePage from "@/components/posts/shirt-posts/post-catalogue.page.vue";
+import EditShirtPage from "@/components/seller-layout/posts/edit-post/edit-shirt.page.vue";
+import PublishShirtPage from "@/components/seller-layout/posts/publish-post/publish-shirt.page.vue";
+import PostCataloguePage from "@/components/seller-layout/posts/shirt-posts/post-catalogue.page.vue";
 export const postsRoutes = [
     { path: '/published', component: PostCataloguePage, meta: { requiresAuth: true, role: 'SELLER' } },
     { path: '/publish', component: PublishShirtPage, meta: { requiresAuth: true, role: 'SELLER' }},
