+++ src/routes/modules/profile.routes.js
@@ -1,8 +1,8 @@
-import EditUserDataPage from "@/components/profile/edit-user-profile/edit-user-data/edit-user-data.page.vue";
-import EditUserPasswordPage from "@/components/profile/edit-user-profile/edit-user-password/edit-user-password.page.vue";
-import EditUserPaymentMethodPage from "@/components/profile/edit-user-profile/edit-user-payment/edit-user-payment-method.page.vue";
-import UserProfilePage from "@/components/profile/user-profile/user-profile.page.vue";
-import EditShirtPage from "@/components/posts/edit-post/edit-shirt.page.vue";
+import EditUserDataPage from "@/components/shared/profile/edit-user-profile/edit-user-data/edit-user-data.page.vue";
+import EditUserPasswordPage from "@/components/shared/profile/edit-user-profile/edit-user-password/edit-user-password.page.vue";
+import EditUserPaymentMethodPage from "@/components/shared/profile/edit-user-profile/edit-user-payment/edit-user-payment-method.page.vue";
+import UserProfilePage from "@/components/shared/profile/user-profile/user-profile.page.vue";
+//import EditShirtPage from "@/components/posts/edit-post/edit-shirt.page.vue";
 
 export const profileRoutes = [
   { path: '/profile', component: UserProfilePage },
