+++ src/routes/modules/shopping-cart.routes.js
@@ -1,4 +1,4 @@
-import UserShoppingCartPage from "@/components/cart/user-shopping-cart/user-shopping-cart.page.vue"
+import UserShoppingCartPage from "@/components/client-layout/cart/user-shopping-cart/user-shopping-cart.page.vue"
 
 export const shoppingCart = [
   { path: '/cart', component: UserShoppingCartPage, meta: { requiresAuth: true, userRole: 'CLIENT' }}
